Added entries for the environment the application is running in (local, test, live) in order to prevent having to manually amend the url's MyCouncil needs to use for the integration modules. Accordingly the integration module entry for emailURl has been expanded.
A version entry has been added so that you can see what version of MyCouncil the page is currently running (it shows in the bottom right of the map). This is useful to identify when users don't refresh the browser after an upgrade and it can still be running the old version after a fix. 

A set of reportItURL items have been added as part of the implementation of the ReportIT module. Likewise an array of reportITentries items have been added which are all the difference categories users will be able to report on.