Message Bar meetings functionality changed to pull in from new JSON file meetings.json (previously the meetings data from contained in the main MyCouncil.json).
Version number now displayed on map (clickable to send the browser to the project page).

Rough copy of the ReportIT module now included. Switch on using the showReportIT json setting. It appears in the main menu when you click on the map. Streetview will show itself if it is available, fill out the form, frame the image in streetview and then click 'Report the problem'. Needs the integration modules installed (not yet released) to go further.